MS-80: refactor select element for sessions form
